+++ fast/stages/02-networking-vpn/spoke-dev.tf
@@ -54,6 +54,7 @@ module "dev-spoke-vpc" {
   name          = "dev-spoke-0"
   mtu           = 1500
   data_folder   = "${var.data_dir}/subnets/dev"
+  psa_ranges    = var.psa_ranges.dev
   subnets_l7ilb = local.l7ilb_subnets.dev
   # set explicit routes for googleapis in case the default route is deleted
   routes = {
@@ -96,17 +97,6 @@ module "dev-spoke-cloudnat" {
   router_network = module.dev-spoke-vpc.name
   router_asn     = 4200001024
   logging_filter = "ERRORS_ONLY"
-}
-
-module "dev-spoke-psa-addresses" {
-  source     = "../../../modules/net-address"
-  project_id = module.dev-spoke-project.project_id
-  psa_addresses = { for r, v in var.psa_ranges.dev : r => {
-    address       = cidrhost(v, 0)
-    network       = module.dev-spoke-vpc.self_link
-    prefix_length = split("/", v)[1]
-    }
-  }
 }
 
 # Create delegated grants for stage3 service accounts