+++ CHANGELOG.md
@@ -12,6 +12,7 @@ All notable changes to this project will be documented in this file.
 - **incompatible change** the variable for service identities IAM has changed in the project factory
 - the `net-vpc` and `project` modules now use the beta provider for shared VPC-related resources
 - add `data-catalog-policy-tag` module
+- **incompatible change** the `psn_ranges` variable has been renamed to `psa_ranges` in the `net-vpc` module and its type changed from `list(string)` to `map(string)`
 
 ## [13.0.0] - 2022-01-27